+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -4,8 +4,8 @@ package com.unciv.build
 object BuildConfig {
     const val kotlinVersion = "1.9.20"
     const val appName = "Unciv"
-    const val appCodeNumber = 943
-    const val appVersion = "4.9.8"
+    const val appCodeNumber = 949
+    const val appVersion = "4.9.13"
 
     const val gdxVersion = "1.12.1"
     const val ktorVersion = "2.2.3"