+++ core/src/com/unciv/logic/civilization/managers/TechManager.kt
@@ -23,7 +23,6 @@ import com.unciv.models.ruleset.unique.UniqueTriggerActivation
 import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.models.ruleset.unit.BaseUnit
 import com.unciv.ui.components.MayaCalendar
-import com.unciv.ui.components.extensions.toPercent
 import com.unciv.ui.components.extensions.withItem
 import kotlin.math.ceil
 import kotlin.math.max
@@ -198,16 +197,8 @@ class TechManager : IsPartOfGameInfoSerialization {
         return (scienceOfLast8Turns.sum() * civInfo.gameInfo.speed.scienceCostModifier).toInt()
     }
 
-    private fun addCurrentScienceToScienceOfLast8Turns() {
-        // The Science the Great Scientist generates does not include Science from Policies, Trade routes and City-States.
-        var allCitiesScience = 0f
-        civInfo.cities.forEach {
-            val totalBaseScience = it.cityStats.baseStatTree.totalStats.science
-            val totalBonusPercents = it.cityStats.statPercentBonusTree.children.asSequence()
-                .filter { it2 -> it2.key != "Policies" }.map { it2 ->  it2.value.totalStats.science }.sum()
-            allCitiesScience += totalBaseScience * totalBonusPercents.toPercent()
-        }
-        scienceOfLast8Turns[civInfo.gameInfo.turns % 8] = allCitiesScience.toInt()
+    private fun addCurrentScienceToScienceOfLast8Turns(science: Int) {
+        scienceOfLast8Turns[civInfo.gameInfo.turns % 8] = science
     }
 
     private fun limitOverflowScience(overflowScience: Int): Int {
@@ -228,7 +219,7 @@ class TechManager : IsPartOfGameInfoSerialization {
     }
 
     fun endTurn(scienceForNewTurn: Int) {
-        addCurrentScienceToScienceOfLast8Turns()
+        addCurrentScienceToScienceOfLast8Turns(scienceForNewTurn)
         if (currentTechnologyName() == null) return
 
         var finalScienceToAdd = scienceForNewTurn
@@ -343,17 +334,25 @@ class TechManager : IsPartOfGameInfoSerialization {
         updateResearchProgress()
     }
 
+    /** A variant of kotlin's [associateBy] that omits null values */
+    private inline fun <T, K, V> Iterable<T>.associateByNotNull(keySelector: (T) -> K, valueTransform: (T) -> V?): Map<K, V> {
+        val destination = LinkedHashMap<K, V>()
+        for (element in this) {
+            val value = valueTransform(element) ?: continue
+            destination[keySelector(element)] = value
+        }
+        return destination
+    }
+
     private fun obsoleteOldUnits(techName: String) {
         // First build a map with obsoleted units to their (nation-specific) upgrade
-        val ruleset = getRuleset()
-        fun BaseUnit.getEquivalentUpgradeOrNull(): BaseUnit? {
-            if (upgradesTo !in ruleset.units) return null  // also excludes upgradesTo==null
-            return civInfo.getEquivalentUnit(upgradesTo!!)
+        fun BaseUnit.getEquivalentUpgradeOrNull(techName: String): BaseUnit? {
+            val unitUpgradesTo = automaticallyUpgradedInProductionToUnitByTech(techName)
+                ?: return null
+            return civInfo.getEquivalentUnit(unitUpgradesTo)
         }
-        val obsoleteUnits = getRuleset().units.asSequence()
-            .filter { it.value.isObsoletedBy(techName) }
-            .map { it.key to it.value.getEquivalentUpgradeOrNull() }
-            .toMap()
+        val obsoleteUnits = getRuleset().units.entries
+            .associateByNotNull({ it.key }, { it.value.getEquivalentUpgradeOrNull(techName) })
         if (obsoleteUnits.isEmpty()) return
 
         // Apply each to all cities - and remember which cities had which obsoleted unit