+++ core/src/com/unciv/models/ruleset/unique/UniqueTriggerActivation.kt
@@ -3,6 +3,7 @@ package com.unciv.models.ruleset.unique
 import com.badlogic.gdx.math.Vector2
 import com.unciv.Constants
 import com.unciv.UncivGame
+import com.unciv.logic.battle.MapUnitCombatant
 import com.unciv.logic.city.City
 import com.unciv.logic.civilization.CivFlags
 import com.unciv.logic.civilization.Civilization
@@ -48,7 +49,10 @@ object UniqueTriggerActivation {
 
         if (!unique.conditionalsApply(civInfo, city)) return false
 
-        val chosenCity = city ?: civInfo.cities.firstOrNull { it.isCapital() }
+        val chosenCity = city ?:
+            tile?.getCity() ?:
+            civInfo.cities.firstOrNull { it.isCapital() }
+
         val tileBasedRandom =
             if (tile != null) Random(tile.position.toString().hashCode())
             else Random(-550) // Very random indeed
@@ -64,20 +68,20 @@ object UniqueTriggerActivation {
 
                 val limit = unit.getMatchingUniques(UniqueType.MaxNumberBuildable)
                     .map { it.params[0].toInt() }.minOrNull()
-                if (limit != null && limit <= civInfo.units.getCivUnits().count { it.name == unitName })
+                if (limit != null && limit <= civInfo.units.getCivUnits().count { it.name == unit.name })
                     return false
 
                 // 4 situations: If city ->
                 val placedUnit = when {
                     city != null || (tile == null && civInfo.cities.isNotEmpty()) ->
-                        civInfo.units.addUnit(unitName, chosenCity) ?: return false
-                    tile != null -> civInfo.units.placeUnitNearTile(tile.position, unitName) ?: return false
+                        civInfo.units.addUnit(unit, chosenCity) ?: return false
+                    tile != null -> civInfo.units.placeUnitNearTile(tile.position, unit) ?: return false
                     civInfo.units.getCivUnits().any() ->
-                        civInfo.units.placeUnitNearTile(civInfo.units.getCivUnits().first().currentTile.position, unitName) ?: return false
+                        civInfo.units.placeUnitNearTile(civInfo.units.getCivUnits().first().currentTile.position, unit) ?: return false
                     else -> return false
                 }
                 val notificationText = getNotificationText(notification, triggerNotificationText,
-                    "Gained [1] [$unitName] unit(s)")
+                    "Gained [1] [${unit.name}] unit(s)")
                     ?: return true
 
                 civInfo.addNotification(
@@ -99,7 +103,7 @@ object UniqueTriggerActivation {
 
                 val limit = unit.getMatchingUniques(UniqueType.MaxNumberBuildable)
                     .map { it.params[0].toInt() }.minOrNull()
-                val unitCount = civInfo.units.getCivUnits().count { it.name == unitName }
+                val unitCount = civInfo.units.getCivUnits().count { it.name == unit.name }
                 val amountFromTriggerable = unique.params[0].toInt()
                 val actualAmount = when {
                     limit == null -> amountFromTriggerable
@@ -111,22 +115,22 @@ object UniqueTriggerActivation {
 
                 val tilesUnitsWerePlacedOn: MutableList<Vector2> = mutableListOf()
                 repeat(actualAmount) {
-                    val placedUnit = if (city != null || tile == null) civInfo.units.addUnit(unitName, chosenCity)
-                        else civInfo.units.placeUnitNearTile(tile.position, unitName)
+                    val placedUnit = if (city != null || tile == null) civInfo.units.addUnit(unit, chosenCity)
+                        else civInfo.units.placeUnitNearTile(tile.position, unit)
                     if (placedUnit != null)
                         tilesUnitsWerePlacedOn.add(placedUnit.getTile().position)
                 }
                 if (tilesUnitsWerePlacedOn.isEmpty()) return true
 
                 val notificationText = getNotificationText(notification, triggerNotificationText,
-                    "Gained [${tilesUnitsWerePlacedOn.size}] [$unitName] unit(s)")
+                    "Gained [${tilesUnitsWerePlacedOn.size}] [${unit.name}] unit(s)")
                     ?: return true
 
                 civInfo.addNotification(
                     notificationText,
                     MapUnitAction(tilesUnitsWerePlacedOn),
                     NotificationCategory.Units,
-                    civInfo.getEquivalentUnit(unit).name
+                    unit.name
                 )
                 return true
             }
@@ -770,6 +774,12 @@ object UniqueTriggerActivation {
                     unit.civ.addNotification(notification, unit.getTile().position, NotificationCategory.Units) // Do we have a heal icon?
                 return true
             }
+            UniqueType.OneTimeUnitDamage -> {
+                MapUnitCombatant(unit).takeDamage(unique.params[0].toInt())
+                if (notification != null)
+                    unit.civ.addNotification(notification, unit.getTile().position, NotificationCategory.Units) // Do we have a heal icon?
+                return true
+            }
             UniqueType.OneTimeUnitGainXP -> {
                 if (!unit.baseUnit.isMilitary()) return false
                 unit.promotions.XP += unique.params[0].toInt()
@@ -802,6 +812,14 @@ object UniqueTriggerActivation {
                     unit.civ.addNotification(notification, unit.getTile().position, NotificationCategory.Units, unit.name)
                 return true
             }
+            UniqueType.OneTimeUnitRemovePromotion -> {
+                val promotion = unit.civ.gameInfo.ruleset.unitPromotions.keys
+                    .firstOrNull { it == unique.params[0]}
+                    ?: return false
+                unit.promotions.removePromotion(promotion)
+                return true
+            }
+
             else -> return triggerCivwideUnique(unique, civInfo = unit.civ, tile=unit.currentTile, triggerNotificationText = triggerNotificationText)
         }
     }